Great work!
You wrote your method without hard coding in the boolean, which is very useful! Great job!!! Sorry I didnt give enough examples in class to get you going on this a little easier! In the top accounts method, you might want to add a system debug so you can check that only 10 print out and that your name change occurs, as expected (easier that going to the the account object and checking the list).  See line 49, where i added it. And just realized it will help me a lot if you dont change the method names as they are written in the homework. That way I can test everyone's homework with the same execute anon statements. I do like your creativity, and can see how that can help to experiment, but it adds another level of admin and error checking for me. Thanks!!