Great job Sara!
I've commented in your name at the top- can you do this with homework as it helps me when I'm running your code. No big deal. Actually, I have to propose a change to your code in order to comment, so that was my excuse to propose a change, but really just to say YAY! So hard to get it all done and I'm glad you did. Also, if you start on something and cant figure it out, dont hesitate to send a link or paste in some homework you are stuck on!